Revert "ipatests: fix circular import for collect_logs"
This reverts commit c45a989506390a4bd159107566b6b2902ede5a30.
